pyproject.toml
[tool.isort]
profile = "black"

[tool.poetry]
name = "campmanager"
version = "1.0.0"
description = ""
authors = ["Ed Rivas"]

[tool.poetry.dependencies]
python = "3.9.4"
Django = "^1.11"
django-grappelli = "2.10.1"
django-import-export = "^0.5"
django-logentry-admin = "^1.0"
django_siterelated = { git = "https://gitlab.com/jerivas/django-siterelated.git", rev = "2ecddfb116923ce2b263b51de7cb2897b9181a5c" }
django_solo = { git = "https://github.com/jerivas/django-solo.git", rev = "30b7b53c0cc894960282be7bbc74a18d69d5df15" }
psycopg2-binary = "^2.8"
xhtml2pdf = "^0.2"
django-environ = "^0.4.5"
gunicorn = "^20.1.0"
whitenoise = {extras = ["brotli"], version = "^5.2.0"}

[tool.poetry.dev-dependencies]
black = "^21.5b0"
django-debug-toolbar = "^1.8"
flake8 = "^3.9.1"
isort = "^5.8.0"
pyupgrade = "^2.14.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
